Revert "Add dependency on Doctrine"
This reverts commit 2bccac836702e6eb8f535009e188af007ef455b1.

Because at this moment we can't use doctrine with the CLI
optimizations because the bundles depend on the HttpKernel component
(or the framework component, I don't recall), so we need to remove
it completely.
At some point we can improve on the CLI optimizations to use the
extensions and copy the bundles extensions to this boilerplate so
that doctrine can be used with all its usual config.

If you are not building a CLI only app, you should remove this
commit and the "Optimize for CLI" commit.